Adding support for updating omega_0 and omega_W values. More below.
Added methods to update the omegas in SIREN. The `SIREN.update_omegas`
method also changes the weights of the network. This is useful when
saving a network after training as a way to remove the omega_0
parameter for inference.

Other methods involve building the network from PTH files and
state_dicts. Also, added a method to reset the weights using
Sitzmann's approach.