Incorporating fast-training changes to the package. More below.
Added the `train_sdf.py` script as a `main.py` replacement in the near
future. It reads the new YAML experiment description file and accepts
parameters to override most of those configurations. This way it
should be easy to have a default experiment file with parameters being
changed in command line using shell scripts.

Also, fixed some issues related to standarization in `meshing.py`, as
well as a torch=>numpy conversion that did not detach the tensor and
moved it to the CPU beforehand.

Added the `PointCloudDeferredSampling` class to the `dataset.py`
file. It reuses most of the code from `PointCloud` while handling
cases where we don't want to sample the SDF values at each iteration
by using the `refresh_sdf` attribute.